Delete duplicate defs of single::FromPublisherOperator
There was one in Single.h and one in SingleOperator.h.  We've been down this
road before with Flowable and Observable...
